rag-backend/requirements.txt
# Flask server
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
pymongo==4.8.0
pypdf==4.3.1

# RAG dependencies
langchain==0.2.11
langchain-community==0.2.11
langchain-google-genai==1.0.8
google-generativeai==0.7.2
sentence-transformers==2.7.0

# Core dependencies
numpy>=1.26.4,<2.0.0
pandas>=2.2.3,<2.4.0
scikit-learn==1.5.1
joblib==1.4.2

# Additional required libs
faiss-cpu==1.8.0
python-dotenv==1.0.1